# Performance improvement of `SparsePauliOp.to_matrix`