<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add a release note file using reno if the change needs to be
  documented in the release notes.
⚠️ If your pull request fixes an open issue, please link to the issue.

- [x] I have added the tests to cover my changes.
- [x] I have updated the documentation accordingly.
- [x] I have read the CONTRIBUTING document.
-->

### Summary

Performance improvement of `SparsePauliOp.to_matrix`.

I used Jake's technique https://github.com/Qiskit/qiskit-terra/pull/9229#pullrequestreview-1202481735.

```python
import random
from timeit import timeit

from qiskit.quantum_info import SparsePauliOp

k = 10
n = 1000

random.seed(123)
op = SparsePauliOp.from_list([(''.join(random.choices('IXYZ', k=k)), 1) for _ in range(n)])

print(f'dense {timeit(lambda: op.to_matrix(sparse=False), number=1)} sec')
print(f'sparse {timeit(lambda: op.to_matrix(sparse=True), number=1)} sec')
```

main
```
dense 9.802448930000082 sec
sparse 3.588473711000006 sec
```

this PR
```
dense 6.786199189000399 sec
sparse 3.1803273659998013 sec
```

### Details and comments

(benchmark for sparse)
```python
import random
from timeit import timeit

from qiskit.quantum_info import SparsePauliOp

k = 25
n = 1

random.seed(123)
op = SparsePauliOp.from_list([(''.join(random.choices('IXYZ', k=k)), 1) for _ in range(n)])

print(f'sparse {timeit(lambda: op.to_matrix(sparse=True), number=1)} sec')
```

main
```
sparse 17.364562793000005 sec
```

this PR
```
sparse 1.894934757000101 sec
```


(further optimization)
If I'm allowed to extend `BasePauli._to_matrix` to multiply `coeff` directly, we can have more improvement for the dense case because Pauli matrices are sparse. See https://github.com/t-imamichi/qiskit-terra/compare/opt-tomatrix2...t-imamichi:qiskit-terra:opt-tomatrix3 for the code diff.

opt-tomatrix3
```
dense 4.616774103999887 sec
sparse 3.021996235000188 sec
```