@t-imamichi 2023-02-22
I added `test_to_matrix_large` to test `to_matrix` method with 10 qubit examples.